% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set}
\alias{set}
\title{Reading is alldata}
\usage{
set()
}
\arguments{
\item{dir}{is directory.}
}
\value{
alldata  #Sys.setenv( JAVA_HOME = 'C:\\Program Files\\Java\\jre1.8.0_151' )
}
\description{
Reading is alldata
}
