% file spatial/man/topo.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{topo}
\alias{topo}
\title{
Spatial Topographic Data
}
\description{
The \code{topo} data frame has 52 rows and 3 columns, of
topographic heights within a 310 feet square.
}
\usage{
topo
}
\format{
This data frame contains the following columns:
\describe{
\item{\code{x}}{
x coordinates (units of 50 feet)
}
\item{\code{y}}{
y coordinates (units of 50 feet)
}
\item{\code{z}}{
heights (feet)
}}}
\source{
Davis, J.C. (1973)
\emph{Statistics and Data Analysis in Geology.}
Wiley.
}
\references{
  Venables, W. N. and Ripley, B. D. (2002)
  \emph{Modern Applied Statistics with S.} Fourth edition.  Springer.
}
\keyword{datasets}
