% file MASS/topo.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{topo}
\alias{topo}
\title{
Spatial Topographic Data
}
\description{
The \code{topo} data frame has 52 rows and 3 columns, of
topographic heights within a 310 feet square.
}
\format{
This data frame contains the following columns:
}
\arguments{
\item{x}{
x coordinates (units of 50 feet)
}
\item{y}{
y coordinates (units of 50 feet)
}
\item{z}{
heights (feet)
}}
\source{
Davis, J.C. (1973) 
\emph{Statistics and Data Analysis in Geology.}
Wiley.
}
\keyword{datasets}
